Add Asset object
Added the most basic representation of an Asset for Latesco to grok.

It includes:

 - A unique identifier
 - Human readable name
 - Symbol for representation

The documentation includes the behavior of Latesco when initializing and
creating a new asset. In addition, the documentation includes explicitly
forbidden behavior of the Asset object. This will ensure the Asset class
stays cohesive.

Finally, tests were implemented to prevent regressions in the future.