Add PriceRecord object
Added the PriceRecord object, which is a thin abstraction over an entry
in the pricing table. It represents an asset's price from a specific API
at a given time.

It includes:

 - Asset UID
 - API UID
 - Timestamp
 - Price

The BigDecimal data structure was used instead of a normal double due to
its more granular control over arithmetic operations, rounding, and
using base 10 representation, which makes it less likely to introduce
error from operations.