Add Api class
Added a class to represent an API. This class is composed of an API UID,
human readable name, domain, and the fetcher.

The class was designed to delegate all fetching of prices to the
ApiFetcher. This allows for Latesco to maintain a contextual history of
prices even if an API is removed in later versions.

In order to support testing, the following dependency was added:

 - Mockito